Add support for complex top-level objects
Previously, json2md only accessed the first element of the top-level
object in the JSON file. That is fine if the JSON file follows a
strictly hierarchical structure (as in the HTML examples), but fails
if there is a more complex structure.
This commit changes this behaviour (at least for the synchronous part)
and now traverses all object keys at the top level, concatenating the
results as we go along.